Overhauling greeter nurse
Trying to overhaul greeter nurse to be able to add multiple patients and send them to the waiting room if they don't have a triage nurse assigned to them yet but are checked in.
Still running into same problems, but I think the overhauled behavior tree might be able to get to a solution after enough debugging.