Greeter nurse overhaul v2
Greeter nurse will now add multiple people to the computer entry table. Need to finish functionality for waiting patients to go to Waiting Room. Need to stress test the overhauled behavior tree.
Also noticed the greeter nurse waits for the triage nurses to be available before entering everyone in, and I can't figure out why, as the parallel node should restart when it hits the fail.