Fixed room problem
The CT's now go to their respective rooms and stay there, but the way I had to accomplish this (there is likely another way but it is slightly more involved) could pose a slight issue in the future. If we allow people to add new agents through some means, if they add a CT, for example, the CT decides which room to go to by its index number being even or odd, so if someone were to add a CT, add something else, and add another CT, both CT's would go to the same room rather than spacing out.